backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pylint>=3.0.0
radon>=6.0.0
